requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
bitarray==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.4
cytoolz==1.0.1
Django==5.2.6
djangorestframework==3.16.1
eth_abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
frozenlist==1.7.0
geographiclib==2.1
geopy==2.4.1
hexbytes==1.3.1
idna==3.10
joblib==1.5.2
multidict==6.6.4
numpy==2.3.3
nvidia-nccl-cu12==2.28.3
packaging==25.0
parsimonious==0.10.0
pillow==11.3.0
pip==24.0
polyline==2.0.3
propcache==0.3.2
py-solc-x==2.0.4
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
pyunormalize==16.0.0
qrcode==8.2
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
scikit-learn==1.7.2
scipy==1.16.2
shapely==2.1.1
sqlparse==0.5.3
threadpoolctl==3.6.0
toolz==1.0.0
types-requests==2.32.4.20250913
typing_extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
web3==7.13.0
websockets==15.0.1
xgboost==2.1.1
yarl==1.20.1
